requirements.txt
-i https://pypi.org/simple
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
fastapi==0.61.0
h11==0.9.0
httptools==0.1.1; sys_platform != 'win32' and sys_platform != 'cygwin' and platform_python_implementation != 'PyPy'
pydantic==1.6.1; python_version >= '3.6'
python-multipart==0.0.5
pyyaml==5.3.1
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
starlette==0.13.6; python_version >= '3.6'
uvicorn==0.11.8
uvloop==0.14.0; sys_platform != 'win32' and sys_platform != 'cygwin' and platform_python_implementation != 'PyPy'
websockets==8.1; python_full_version >= '3.6.1'
gym>=0.14
numpy>=1.17.0
scikit-image>=0.14.5
keras>=2.3.0
Pillow
pygame>=1.9.6
tqdm
plotly
